refactor: FillingPaintTool 로 중복 제거
